Refactored how the agent is constructed to a utility function.  Moved the ReasoningModelContext to autogen_utils.py.
